server/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-MySQLdb==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
PyJWT==1.7.1
Werkzeug==1.0.1
